refactor: split p2p event handling
